requirements.txt
Django==1.6.4
MySQL-python==1.2.5
South==0.8.4
django-filebrowser==3.5.6
django-grappelli==2.5.3
Pillow==2.4.0
bpython==0.13
python-dateutil==2.2
flup==1.0.2
django-debug-toolbar==1.2.1
